improve error handling for switch delete
there's a lot here, let's break it down.

this uses the richer error model for gRPC described here:

https://grpc.io/docs/guides/error/#richer-error-model

and here:

https://google.aip.dev/193#error_model

and using examples from:

https://github.com/grpc/grpc-go/tree/master/examples/features/error_details

There's a lot more of this kind of thing to be done in the gRPC api.
This is just a first example.

I updated golangci-lint and had to cope with the fallout from that,
including a lot of gosec related cast issues.Fixing those also lead to
forcing all TCP port values to be uint16 as it seems they should be,
keeping disk/iso file sizes as uint64 and some other cleanups.

Having the newer golangci-lint meant the ability to add
google.golang.org/grpc/internal/status to ignorePackageGlobs which was
needed for the gRPC error handling changes.

+++ cirrinad/config/config.go
@@ -29,8 +29,8 @@ type Info struct {
 		Grpc struct {
 			// TODO separate settings for IPv4 and IPv6 IP
 			IP      string
-			Port    uint
-			Timeout uint64 `default:"60"` // in seconds
+			Port    uint16
+			Timeout int64 `default:"60"` // in seconds
 		}
 		Mac struct {
 			Oui string
@@ -45,17 +45,17 @@ type Info struct {
 	Vnc struct {
 		// TODO separate settings for IPv4 and IPv6 IP
 		IP   string `default:"0.0.0.0"`
-		Port uint   `default:"5900"`
+		Port uint16 `default:"5900"`
 	}
 	Debug struct {
 		// TODO separate settings for IPv4 and IPv6 IP
 		IP   string `default:"0.0.0.0"`
-		Port uint   `default:"2828"`
+		Port uint16 `default:"2828"`
 	}
 	Metrics struct {
 		Enabled bool   `default:"false"`
 		Host    string `default:""`
-		Port    uint   `default:"2223"`
+		Port    uint16 `default:"2223"`
 	}
 }
